# crates/api/src/routes/families.rs
//! `/families` CRUD + `/families/{id}/invites` endpoints.
//!
//! - `GET    /families/me`             — list every family the caller belongs
//!   to. Reads memberships fresh from the DB (not the JWT) so it can surface
//!   each family's `created_at` for the switcher without bloating the token.
//! - `POST   /families`                — create a new family, auto-join the
//!   creator as `Owner`, then reissue a fresh access cookie so the new
//!   membership is immediately reflected in the JWT.
//! - `PATCH  /families/{id}`           — rename. Requires `Admin` or `Owner`
//!   on the target family.
//! - `DELETE /families/{id}`           — delete. Requires `Owner` on the
//!   target family. Returns `{ "data": null }` per spec §5.
//! - `POST   /families/{id}/invites`   — issue a single-use invite token,
//!   email it. Requires `Admin` or `Owner`. Cannot invite as `Owner`.
//!
//! Authorisation pattern: every per-family handler resolves the target id
//! against `claims.all_families` first (returning `NotFamilyMember` when the
//! caller isn't in the token), then [`crate::auth::require_role`] against the
//! resolved membership. We deliberately do NOT require the `X-Family-Id`
//! header here: these handlers identify their target via the path segment,
//! and the header is reserved for handlers whose target family is implicit
//! (persons, contacts, reminders).

use actix_web::{HttpRequest, HttpResponse, delete, get, patch, post, web};
use chrono::{DateTime, Duration, Utc};
use my_fam_tree_domain::{FamilyId, Role};
use my_fam_tree_email::{Locale as EmailLocale, render_invite};
use serde::{Deserialize, Serialize};
use utoipa::ToSchema;
use uuid::Uuid;

use crate::auth::{ActiveFamily, generate_opaque_token};
use crate::cookies::access_cookie;
use crate::response::NullResponseBody;
use crate::services::audit;
use crate::services::auth_service::issue_access_token_for;
use crate::validation::{
    email_invalid, looks_like_email, role_invalid, string_too_long, value_required,
};

// Family-name cap (security audit MEDIUM). The same value flows verbatim
// into the invite + owner-transfer email Subject lines (`templates.rs`);
// lettre escapes CR/LF in headers, so this is the second line of defense
// against header-injection on the off-chance the template path changes.
const FAMILY_NAME_MAX: u32 = 200;

/// Trim and validate a family name. Rejects empty, over-long, or
/// strings containing CR/LF (header-injection hardening).
fn validate_family_name(raw: &str) -> Result<String, ApiError> {
    let trimmed = raw.trim();
    if trimmed.is_empty() {
        return Err(value_required("/name"));
    }
    if trimmed.contains('\r') || trimmed.contains('\n') {
        return Err(ApiError::Validation(vec![crate::error::FieldViolation::new(
            "/name",
            "validation.value_invalid",
            "family name must not contain newline characters",
        )]));
    }
    if u32::try_from(trimmed.chars().count()).unwrap_or(u32::MAX) > FAMILY_NAME_MAX {
        return Err(string_too_long("/name", FAMILY_NAME_MAX));
    }
    Ok(trimmed.to_string())
}
use crate::{ApiError, ApiResponse, AppState, response_body};

// ---------------------------------------------------------------------------
// Request / response DTOs.
// ---------------------------------------------------------------------------

#[derive(Debug, Serialize, ToSchema)]
pub struct FamilyView {
    pub id: Uuid,
    pub name: String,
    pub role: Role,
    /// Family creation time (RFC 3339). Present on `GET /families/me` so the
    /// switcher can disambiguate same-named families by date; `None` on the
    /// JWT-projection paths (rename, invite-accept) that don't carry it.
    #[serde(skip_serializing_if = "Option::is_none")]
    pub created_at: Option<DateTime<Utc>>,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct MyFamiliesRes {
    pub families: Vec<FamilyView>,
}

#[derive(Debug, Deserialize, ToSchema)]
#[serde(deny_unknown_fields)]
pub struct CreateFamilyReq {
    pub name: String,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct CreateFamilyRes {
    pub family: FamilyView,
    pub claims: super::auth::ConsumeRes,
}

#[derive(Debug, Deserialize, ToSchema)]
#[serde(deny_unknown_fields)]
pub struct RenameFamilyReq {
    pub name: String,
}

#[derive(Debug, Deserialize, ToSchema)]
#[serde(deny_unknown_fields)]
pub struct InviteReq {
    pub email: String,
    pub role: Role,
    /// Optional person row this invite is bound to. On accept, the API
    /// atomically sets `persons.linked_user_id = new_user.id` so the
    /// recipient becomes the person they were invited as. Nullable: an
    /// admin may invite without binding to a specific person.
    #[serde(default)]
    pub person_id: Option<Uuid>,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct InviteRes {
    pub status: &'static str,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct LatestPersonView {
    pub id: Uuid,
    pub given_name: String,
    pub family_name: String,
    pub created_at: DateTime<Utc>,
}

/// Aggregated stats for the admin "Family" overview page.
///
/// Single round-trip so the page doesn't need to fan out to three
/// separate endpoints (count members, count persons, fetch latest 3
/// persons). Returned by `GET /admin/family/overview`.
#[derive(Debug, Serialize, ToSchema)]
pub struct FamilyOverview {
    pub id: Uuid,
    pub name: String,
    pub role: Role,
    pub member_count: u64,
    pub person_count: u64,
    /// At most 3 entries, ordered newest-first by `created_at`.
    pub latest_persons: Vec<LatestPersonView>,
}

response_body!(pub MyFamiliesResponseBody, MyFamiliesRes);
response_body!(pub CreateFamilyResponseBody, CreateFamilyRes);
response_body!(pub FamilyViewResponseBody, FamilyView);
response_body!(pub InviteResponseBody, InviteRes);
response_body!(pub FamilyOverviewResponseBody, FamilyOverview);

// ---------------------------------------------------------------------------
// Helpers.
// ---------------------------------------------------------------------------

/// Find `family_id` among the caller's JWT memberships, returning a thin
/// `ActiveFamily` snapshot the role checker can consume.
pub(crate) fn resolve_membership(
    claims: &crate::auth::UserClaims,
    family_id: Uuid,
) -> Result<ActiveFamily, ApiError> {
    claims
        .all_families
        .iter()
        .find(|f| f.id.into_uuid() == family_id)
        .map(|f| ActiveFamily { id: f.id, name: f.name.clone(), role: f.role })
        .ok_or(ApiError::NotFamilyMember(family_id))
}

fn seconds_i64(value: u64) -> i64 {
    i64::try_from(value).unwrap_or(i64::MAX)
}

fn internal<E: std::fmt::Display>(e: E) -> ApiError {
    ApiError::Internal(anyhow::anyhow!(e.to_string()))
}

// ---------------------------------------------------------------------------
// GET /families/me
// ---------------------------------------------------------------------------

#[utoipa::path(
    get,
    path = "/api/v1/families/me",
    responses(
        (status = 200, description = "Caller's family memberships", body = MyFamiliesResponseBody),
        (status = 401, description = "No session"),
    ),
    security(("cookie_access" = [])),
    tag = "families",
)]
#[allow(clippy::future_not_send)]
#[allow(unreachable_pub)]
#[get("/families/me")]
pub async fn list_mine(
    state: web::Data<AppState>,
    req: HttpRequest,
) -> Result<ApiResponse<MyFamiliesRes>, ApiError> {
    let claims = crate::auth::user_claims(&req)?;
    // Fresh DB read (not the JWT claim) so we can surface `created_at` for the
    // switcher without bloating the signed token. Reflects current memberships.
    let rows = state.memberships.list_for_user(claims.user_id).await.map_err(internal)?;
    let families = rows
        .into_iter()
        .map(|m| FamilyView {
            id: m.family_id.into_uuid(),
            name: m.family_name,
            role: m.role,
            created_at: Some(m.created_at),
        })
        .collect();
    Ok(ApiResponse::ok(MyFamiliesRes { families }))
}

// ---------------------------------------------------------------------------
// POST /families
// ---------------------------------------------------------------------------

#[utoipa::path(
    post,
    path = "/api/v1/families",
    operation_id = "families_create",
    request_body = CreateFamilyReq,
    responses(
        (status = 200, description = "Family created and caller auto-joined as Owner", body = CreateFamilyResponseBody),
        (status = 401, description = "No session"),
        (status = 422, description = "Validation failed"),
    ),
    security(("cookie_access" = [])),
    tag = "families",
)]
#[allow(clippy::future_not_send)]
#[allow(unreachable_pub)]
#[post("/families")]
pub async fn create(
    state: web::Data<AppState>,
    req: HttpRequest,
    body: web::Json<CreateFamilyReq>,
) -> Result<HttpResponse, ApiError> {
    let claims = crate::auth::user_claims(&req)?;
    let name = validate_family_name(&body.name)?;

    let family = state.families.create(&name, claims.user_id).await.map_err(internal)?;
    state.memberships.insert(family.id, claims.user_id, Role::Owner).await.map_err(internal)?;
    audit::record(
        &state.audit,
        family.id,
        claims.user_id,
        "create",
        "membership",
        None,
        serde_json::json!({
            "user_id": claims.user_id.into_uuid(),
            "role": "owner",
        }),
    )
    .await;

    // Reissue access JWT so the new membership is visible immediately.
    let user = state
        .users
        .find_by_id(claims.user_id)
        .await
        .map_err(internal)?
        .ok_or(ApiError::Unauthenticated)?;
    let (access, fams) = issue_access_token_for(&state.jwt_issuer, &state.memberships, &user)
        .await
        .map_err(ApiError::Internal)?;

    let claims_payload = super::auth::ConsumeRes {
        user_id: user.id.into_uuid(),
        email: user.email.clone(),
        locale: user.locale.as_str().to_string(),
        families: fams,
    };
    let response = CreateFamilyRes {
        family: FamilyView {
            id: family.id.into_uuid(),
            created_at: Some(family.created_at),
            name: family.name,
            role: Role::Owner,
        },
        claims: claims_payload,
    };

    let mut resp = HttpResponse::Ok().json(ApiResponse::ok(response));
    let _ = resp.add_cookie(&access_cookie(&state.cfg, access));
    Ok(resp)
}

// ---------------------------------------------------------------------------
// PATCH /families/{id}
// ---------------------------------------------------------------------------

#[utoipa::path(
    patch,
    path = "/api/v1/families/{id}",
    request_body = RenameFamilyReq,
    responses(
        (status = 200, description = "Family renamed", body = FamilyViewResponseBody),
        (status = 401, description = "No session"),
        (status = 403, description = "Insufficient role"),
        (status = 422, description = "Validation failed"),
    ),
    security(("cookie_access" = [])),
    tag = "families",
)]
#[allow(clippy::future_not_send)]
#[allow(unreachable_pub)]
#[patch("/families/{id}")]
pub async fn rename(
    state: web::Data<AppState>,
    req: HttpRequest,
    path: web::Path<Uuid>,
    body: web::Json<RenameFamilyReq>,
) -> Result<ApiResponse<FamilyView>, ApiError> {
    let claims = crate::auth::user_claims(&req)?;
    let id = path.into_inner();
    let active = resolve_membership(&claims, id)?;
    crate::auth::require_db_role(&state, claims.user_id, active.id, Role::Admin).await?;

    let name = validate_family_name(&body.name)?;
    state.families.rename(FamilyId::from_uuid(id), &name).await.map_err(internal)?;
    Ok(ApiResponse::ok(FamilyView { id, name, role: active.role, created_at: None }))
}

// ---------------------------------------------------------------------------
// GET /admin/family/overview
// ---------------------------------------------------------------------------

#[utoipa::path(
    get,
    path = "/api/v1/admin/family/overview",
    responses(
        (status = 200, description = "Active family overview (admin+owner only)", body = FamilyOverviewResponseBody),
        (status = 401, description = "No session"),
        (status = 403, description = "Insufficient role"),
        (status = 422, description = "Missing X-Family-Id header"),
    ),
    security(("cookie_access" = [])),
    tag = "families",
)]
#[allow(clippy::future_not_send)]
#[allow(unreachable_pub)]
#[get("/admin/family/overview")]
pub async fn admin_overview(
    state: web::Data<AppState>,
    req: HttpRequest,
) -> Result<ApiResponse<FamilyOverview>, ApiError> {
    // Active family resolution: takes X-Family-Id from the request and
    // matches it against the caller's JWT memberships. Then the DB role
    // check makes sure they're admin / owner — the route guard on the
    // FE also enforces this, but the server is the source of truth.
    let (claims, active) = crate::auth::user_claims_with_family(&req)?;
    crate::auth::require_db_role(&state, claims.user_id, active.id, Role::Admin).await?;

    // The membership check above already proves the family exists in the
    // caller's claims; a `None` here would be an internal consistency
    // bug (membership row pointing at a deleted family), not a
    // user-facing 404.
    let family =
        state.families.find_by_id(active.id).await.map_err(internal)?.ok_or_else(|| {
            ApiError::Internal(anyhow::anyhow!("family vanished for active membership"))
        })?;
    // Three aggregations against the same family. Parallel awaits keep
    // the cumulative latency tight; the page is one of those "load
    // everything at once" surfaces, not a hot path.
    let (members, persons, latest) = tokio::join!(
        state.memberships.count_in_family(active.id),
        state.persons.count_in_family(active.id),
        state.persons.list_latest_in_family(active.id, 3),
    );
    let member_count = members.map_err(internal)?;
    let person_count = persons.map_err(internal)?;
    let latest_persons = latest
        .map_err(internal)?
        .into_iter()
        .map(|p| LatestPersonView {
            id: p.id.into_uuid(),
            given_name: p.given_name,
            family_name: p.family_name,
            created_at: p.created_at,
        })
        .collect();
    Ok(ApiResponse::ok(FamilyOverview {
        id: active.id.into_uuid(),
        name: family.name,
        role: active.role,
        member_count,
        person_count,
        latest_persons,
    }))
}

// ---------------------------------------------------------------------------
// DELETE /families/{id}
// ---------------------------------------------------------------------------

#[utoipa::path(
    delete,
    path = "/api/v1/families/{id}",
    responses(
        (status = 200, description = "Family deleted", body = NullResponseBody, content_type = "application/json"),
        (status = 401, description = "No session"),
        (status = 403, description = "Insufficient role"),
    ),
    security(("cookie_access" = [])),
    tag = "families",
)]
#[allow(clippy::future_not_send)]
#[allow(unreachable_pub)]
#[delete("/families/{id}")]
pub async fn delete_family(
    state: web::Data<AppState>,
    req: HttpRequest,
    path: web::Path<Uuid>,
) -> Result<ApiResponse<serde_json::Value>, ApiError> {
    let claims = crate::auth::user_claims(&req)?;
    let id = path.into_inner();
    let active = resolve_membership(&claims, id)?;
    crate::auth::require_db_role(&state, claims.user_id, active.id, Role::Owner).await?;
    state.families.delete(FamilyId::from_uuid(id)).await.map_err(internal)?;
    Ok(ApiResponse::ok(serde_json::Value::Null))
}

// ---------------------------------------------------------------------------
// POST /families/{id}/invites
// ---------------------------------------------------------------------------

#[utoipa::path(
    post,
    path = "/api/v1/families/{id}/invites",
    request_body = InviteReq,
    responses(
        (status = 200, description = "Invite created and email queued", body = InviteResponseBody),
        (status = 401, description = "No session"),
        (status = 403, description = "Insufficient role"),
        (status = 422, description = "Validation failed"),
    ),
    security(("cookie_access" = [])),
    tag = "families",
)]
#[allow(clippy::future_not_send)]
#[allow(unreachable_pub)]
#[post("/families/{id}/invites")]
pub async fn invite(
    state: web::Data<AppState>,
    req: HttpRequest,
    path: web::Path<Uuid>,
    body: web::Json<InviteReq>,
) -> Result<ApiResponse<InviteRes>, ApiError> {
    let claims = crate::auth::user_claims(&req)?;
    let family_id = path.into_inner();
    let active = resolve_membership(&claims, family_id)?;
    crate::auth::require_db_role(&state, claims.user_id, active.id, Role::Admin).await?;

    if body.role == Role::Owner {
        return Err(role_invalid("/role", "cannot invite as owner"));
    }

    let email = body.email.trim().to_lowercase();
    if !looks_like_email(&email) {
        return Err(email_invalid("/email"));
    }

    // Reject duplicates BEFORE generating a token / sending email. Two
    // pending invites for the same email in the same family would race on
    // accept (only one can match the freshly-signed-in JWT email) and
    // pollute the admin pending-list with stale rows.
    let fid = FamilyId::from_uuid(family_id);
    if state.invites.find_pending_by_email(fid, &email).await.map_err(internal)?.is_some() {
        return Err(ApiError::InviteDuplicate);
    }

    // Cross-family IDOR guard (security audit MEDIUM). `family_invites.person_id`
    // only `REFERENCES persons(id)`, not the invite's family_id — an F1
    // admin could otherwise mint an invite referencing an F2 person. On
    // accept the linkage step would then fail (404 mapped to 500) but the
    // membership row + audit "verify" event would already have landed,
    // leaving a member without the intended linkage. Reject at create time.
    if let Some(pid) = body.person_id
        && state
            .persons
            .find_in_family(fid, my_fam_tree_domain::PersonId::from_uuid(pid))
            .await
            .map_err(internal)?
            .is_none()
    {
        return Err(ApiError::PersonNotFound { id: Some(pid) });
    }

    let (token, hash) = generate_opaque_token();
    let invite_id = state
        .invites
        .create(
            fid,
            &email,
            body.role,
            claims.user_id,
            body.person_id,
            &hash,
            Utc::now() + Duration::seconds(seconds_i64(state.cfg.magic_link.invite_ttl_seconds)),
        )
        .await
        .map_err(internal)?;

    // Locale-prefixed for the inviter's language (the invitee may not have an
    // account yet, so the inviter's preference is the best signal we have).
    let locale = EmailLocale::from_str_or_en(&claims.locale);
    let link = locale.link(&state.cfg.web.public_url, &format!("/invite/accept?token={token}"));
    let email_msg =
        render_invite(locale, &state.cfg.web.public_url, &active.name, &claims.email, &link)
            .map_err(internal)?;
    // Enqueue into the durable outbox — the worker drains via SMTP. Same
    // request thread used to block on SmtpSender.send(); now it returns as
    // soon as Postgres has the row.
    state
        .outbox
        .enqueue(&my_fam_tree_domain::EmailOutboxInsert {
            kind: my_fam_tree_domain::EmailOutboxKind::INVITE.to_string(),
            to_addr: email.clone(),
            subject: email_msg.subject,
            text_body: email_msg.text,
            html_body: Some(email_msg.html),
        })
        .await
        .map_err(internal)?;

    audit::record(
        &state.audit,
        fid,
        claims.user_id,
        "invite",
        "membership",
        Some(invite_id),
        serde_json::json!({
            "email": email,
            "role": body.role,
            "person_id": body.person_id,
        }),
    )
    .await;
    Ok(ApiResponse::ok(InviteRes { status: "sent" }))
}

// ---------------------------------------------------------------------------
// Tests — pure-logic helpers; HTTP integration tests live in tests/.
// ---------------------------------------------------------------------------

#[cfg(test)]
#[allow(
    clippy::unwrap_used,
    clippy::expect_used,
    clippy::indexing_slicing,
    clippy::panic,
    clippy::future_not_send
)]
mod tests {
    use my_fam_tree_domain::{FamilyId, Role, UserId};
    use uuid::Uuid;

    use super::*;
    use crate::auth::user_claims::{FamilyMembershipMirror, UserClaims};

    fn mk_claims(memberships: Vec<(Uuid, &'static str, Role)>) -> UserClaims {
        UserClaims {
            user_id: UserId::from_uuid(Uuid::new_v4()),
            email: "a@b.c".into(),
            locale: "en".into(),
            active_family: None,
            all_families: memberships
                .into_iter()
                .map(|(id, name, role)| FamilyMembershipMirror {
                    id: FamilyId::from_uuid(id),
                    name: name.into(),
                    role,
                })
                .collect(),
        }
    }

    #[test]
    fn resolve_membership_returns_active_family_when_present() {
        let fid = Uuid::new_v4();
        let claims = mk_claims(vec![(fid, "Müller", Role::Owner)]);
        let active = resolve_membership(&claims, fid).expect("resolved");
        assert_eq!(active.id.into_uuid(), fid);
        assert_eq!(active.name, "Müller");
        assert_eq!(active.role, Role::Owner);
    }

    #[test]
    fn resolve_membership_errors_when_not_in_token() {
        let claims = mk_claims(vec![(Uuid::new_v4(), "Müller", Role::Owner)]);
        let missing = Uuid::new_v4();
        match resolve_membership(&claims, missing) {
            Err(ApiError::NotFamilyMember(id)) => assert_eq!(id, missing),
            _ => panic!("expected NotFamilyMember"),
        }
    }

    #[test]
    fn seconds_i64_clamps_overflow_to_max() {
        assert_eq!(seconds_i64(0), 0);
        assert_eq!(seconds_i64(900), 900);
        assert_eq!(seconds_i64(u64::MAX), i64::MAX);
    }

    #[test]
    fn invite_res_serialises_with_status_field() {
        let v = serde_json::to_value(InviteRes { status: "sent" }).unwrap();
        assert_eq!(v["status"], "sent");
    }

    #[test]
    fn create_family_req_deserialises_from_name_field() {
        let r: CreateFamilyReq =
            serde_json::from_value(serde_json::json!({"name": "Müller"})).unwrap();
        assert_eq!(r.name, "Müller");
    }

    #[test]
    fn invite_req_deserialises_email_and_role() {
        let r: InviteReq =
            serde_json::from_value(serde_json::json!({"email": "a@b.co", "role": "admin"}))
                .unwrap();
        assert_eq!(r.email, "a@b.co");
        assert_eq!(r.role, Role::Admin);
        assert!(r.person_id.is_none());
    }

    #[test]
    fn invite_req_deserialises_with_person_id() {
        let pid = Uuid::new_v4();
        let r: InviteReq = serde_json::from_value(
            serde_json::json!({"email": "a@b.co", "role": "user", "person_id": pid}),
        )
        .unwrap();
        assert_eq!(r.person_id, Some(pid));
    }

    #[test]
    fn family_view_serialises_with_role_as_string() {
        let v = serde_json::to_value(FamilyView {
            id: Uuid::nil(),
            name: "Müller".into(),
            role: Role::Owner,
            created_at: None,
        })
        .unwrap();
        assert_eq!(v["name"], "Müller");
        assert!(v.get("role").is_some());
    }
}