# fe/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "my-fam-tree API",
    "description": "Family platform API",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.18"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "paths": {
    "/api/v1/admin/family/overview": {
      "get": {
        "tags": [
          "families"
        ],
        "operationId": "admin_overview",
        "responses": {
          "200": {
            "description": "Active family overview (admin+owner only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyOverviewResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "403": {
            "description": "Insufficient role"
          },
          "422": {
            "description": "Missing X-Family-Id header"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
    "/api/v1/auth/consume": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "consume",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Magic link consumed, session established",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumeResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Magic link invalid or expired"
          }
        }
      }
    },
...
    "/api/v1/auth/magic-link": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "magic_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Magic link sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicLinkResponseBody"
                }
              }
            }
          },
          "422": {
            "description": "Invalid email"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/v1/families": {
      "post": {
        "tags": [
          "families"
        ],
        "operationId": "families_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFamilyReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Family created and caller auto-joined as Owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFamilyResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "422": {
            "description": "Validation failed"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
    "/api/v1/families/me": {
      "get": {
        "tags": [
          "families"
        ],
        "operationId": "list_mine",
        "responses": {
          "200": {
            "description": "Caller's family memberships",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyFamiliesResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
...
    "/api/v1/families/{id}": {
      "delete": {
        "tags": [
          "families"
        ],
        "operationId": "delete_family",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Family deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "patch": {
        "tags": [
          "families"
        ],
        "operationId": "rename",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFamilyReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Family renamed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyViewResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "403": {
            "description": "Insufficient role"
          },
          "422": {
            "description": "Validation failed"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
    "/api/v1/families/{id}/invites": {
      "get": {
        "tags": [
          "invites"
        ],
        "operationId": "invites_list_pending",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Family id (must be a family the caller belongs to)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending invites for this family",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitesListResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "403": {
            "description": "Insufficient role (admin / owner required)"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "post": {
        "tags": [
          "families"
        ],
        "operationId": "invite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Invite created and email queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "403": {
            "description": "Insufficient role"
          },
          "422": {
            "description": "Validation failed"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsumeReq": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ConsumeResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConsumeRes"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "CreateFamilyReq": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CreateFamilyResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateFamilyRes"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "FamilyOverviewResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FamilyOverview"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "MagicLinkReq": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "locale": {
            "type": [
              "string",
              "null"
            ],
            "description": "Locale the FE detected on the sign-in page (`\"en\"` / `\"de\"`). Optional\nfor backwards-compatible callers; used to seed a brand-new account's\npreference AND to locale-prefix the magic-link URL. For a RETURNING\nuser their stored account locale wins (it's their explicit setting),\nso this only takes effect when the account doesn't exist yet."
          }
        },
        "additionalProperties": false
      },
...
      "MagicLinkResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MagicLinkRes"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookie_access": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access"
      }
    }
  },
...
}